8.NoSQL/300150295/requirements.txt
psycopg2-binary
